chore: migrate common set integration tests
