+++ packages/client-sdk-nodejs/test/integration/shared/set.test.ts
@@ -1,0 +1,6 @@
+import {runSetTests} from '@gomomento/common-integration-tests';
+import {SetupIntegrationTest} from '../integration-setup';
+
+const {Momento, IntegrationTestCacheName} = SetupIntegrationTest();
+
+runSetTests(Momento, IntegrationTestCacheName);
