+++ packages/common-integration-tests/src/index.ts
@@ -4,3 +4,4 @@ export * from './get-set-delete';
 export * from './generate-api-token';
 export * from './list';
 export * from './dictionary';
+export * from './set';
